test: split new demangling tests into their own module
Codex review: this PR's additions pushed tests/test_sprint9_html.py to
1212 lines, over the architecture gate's 1200-line test-module cap
(scripts/check_architecture.py's new-test-size check), with no adoption
debt entry to allow it.

Moved TestRealDemanglingThroughTheProductionChangeDataclass (the whole
self-contained block this PR added, 348 lines) into a new
tests/test_html_report_demangle.py, leaving test_sprint9_html.py at
864 lines -- back under the parent's original 851-line size plus the
one small pre-existing test_write_passes_demangle_through addition.

Verified: same 76 tests pass across both files post-split as before
it; scripts/check_architecture.py now reports 0 errors.
